Icon overhaul
Changed icons to use lucide icons in addition to tabler icons. Removed CDN imports for offline support (except nerdfonts CDN, which is used for weather icons, which don't show when there's no internet anyway).

In order to achieve treeshaking, the iconMap json had to be swapped with a static import list (if you know a better way to do this, please open a pr)